Skip preorder items
Preorder items are not guaranteed to be complete (eg. a preordered album may only have 3 out of 10 tracks available); however at present they are not synced correctly when they are released, as the `bandcamp_item_id.txt` file that is created when initially syncing prevents this. We avoid this by skipping items marked as preorders entirely.